--- common/src/main/java/com/alibaba/nacos/common/utils/BiFunction.java
@@ -1,38 +1,0 @@
-/*
- * Copyright 1999-2018 Alibaba Group Holding Ltd.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.alibaba.nacos.common.utils;
-
-/**
- * Represents a function that accepts two arguments and produces a result. The following utility functions are extracted
- * from org.apache.commons.lang3.
- *
- * <p>This is a <a href="package-summary.html">functional interface</a>
- * whose functional method is {@link #apply(Object, Object)}.
- *
- * @author zongtanghu
- */
-public interface BiFunction<T, U, R> {
-    
-    /**
-     * Applies this function to the two given arguments.
-     *
-     * @param t the first function argument
-     * @param u the second function argument
-     * @return the function result
-     */
-    R apply(T t, U u);
-}
