+++ common/src/main/java/com/alibaba/nacos/common/utils/MapUtil.java
@@ -23,6 +23,7 @@ import java.util.Dictionary;
 import java.util.Map;
 import java.util.Objects;
 import java.util.function.Predicate;
+import java.util.function.BiFunction;
 
 /**
  * Map utils.
