+++ naming/src/main/java/com/alibaba/nacos/naming/selector/context/NoneSelectorContextBuilder.java
@@ -15,39 +15,27 @@
  *
  */
 
-package com.alibaba.nacos.naming.selector;
+package com.alibaba.nacos.naming.selector.context;
 
-import com.alibaba.nacos.api.exception.NacosException;
 import com.alibaba.nacos.api.naming.pojo.Instance;
-import com.alibaba.nacos.api.selector.Selector;
+import com.alibaba.nacos.api.selector.context.SelectorContextBuilder;
 
 import java.util.List;
 
 /**
- * Selector with no filtering. The logic is same to {@link com.alibaba.nacos.naming.selector.v1.NoneSelector}.
+ * The {@link NoneSelectorContextBuilder} will return the provider as context for the {@link com.alibaba.nacos.api.selector.Selector}
+ * which doesn't need any other resource.
  *
  * @author chenglu
- * @date 2021-08-04 13:28
+ * @date 2021-08-04 13:31
  */
-public class NoneSelector<T extends Instance> implements Selector<List<T>, List<T>, String> {
+public class NoneSelectorContextBuilder<T extends Instance>  implements SelectorContextBuilder<List<T>, String, List<T>> {
     
     private static final String CONTEXT_TYPE = "NONE";
     
-    private static final String TYPE = "none";
-    
     @Override
-    public Selector<List<T>, List<T>, String> parse(String condition) throws NacosException {
-        return this;
-    }
-    
-    @Override
-    public List<T> select(List<T> context) {
-        return context;
-    }
-    
-    @Override
-    public String getType() {
-        return TYPE;
+    public List<T> build(String consumer, List<T> provider) {
+        return provider;
     }
     
     @Override
