+++ naming/src/test/resources/META-INF/services/com.alibaba.nacos.api.selector.Selector
